Added ability to create and refresh JWT token
Added the ability to create JWT token.

Supports custom token as base key if necessary. Base key is the signing key
that is required to sign and verify the JWT token. If the base key is present,
then the JWT token set will use the base key. For other scenarios, the base key
is generated by the library itself.

Added the ability to refresh JWT token.

Support refresh of JWT token if the validity is within 30 seconds of expiry. This is
set forth as a hard rule and is not customizable by default. If the validity time is
within the 30 seconds mark, then the JWT is resigned with extra time that the client
requested with.

Added unit test

Added and updated unit tests to verify main methods and util methods that were used
during this package. Coverage is at 80%.